+++ IDatabase/IDatabase.h
@@ -9,8 +9,8 @@ public:
 
     virtual ~IDatabase() {}
 
-    virtual std::vector<int> Select(int id) = 0;
-    virtual bool Insert(int id, const std::vector<int>&) = 0;
+    virtual std::vector<size_t> Select(int id) = 0;
+    virtual bool Insert(int id, const std::vector<size_t>&) = 0;
     virtual bool Insert(Result a) = 0;
 };
